refactor: cleaned up API v1 and v2
