+++ Lavinia-api/Data/NOContext.cs
@@ -3,7 +3,6 @@ using Microsoft.EntityFrameworkCore;
 
 namespace LaviniaApi.Data
 {
-    // API v2
     public class NOContext : DbContext
     {
         public NOContext()
@@ -17,7 +16,6 @@ namespace LaviniaApi.Data
         public DbSet<PartyVotes> PartyVotes { get; set; }
         public DbSet<DistrictMetrics> DistrictMetrics { get; set; }
         public DbSet<ElectionParameters> ElectionParameters { get; set; }
-        public DbSet<ElectionParametersV3> ElectionParametersV3 { get; set; }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
@@ -28,8 +26,6 @@ namespace LaviniaApi.Data
             modelBuilder.Entity<DistrictMetrics>()
                 .HasKey(d => new {d.ElectionYear, d.District});
             modelBuilder.Entity<ElectionParameters>()
-                .HasKey(e => new {e.ElectionYear, e.ElectionType});
-            modelBuilder.Entity<ElectionParametersV3>()
                 .HasKey(e => new { e.ElectionYear, e.ElectionType });
         }
     }
