+++ Lavinia-api/Models/Party.cs
@@ -3,7 +3,6 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace LaviniaApi.Models
 {
-    // API v3
     // Stores all the parties in the API
     public class Party
     {
