--- Lavinia-api/Models/Result.cs
@@ -1,16 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations.Schema;
-
-namespace LaviniaApi.Models
-{
-    // API v1
-    public class Result
-    {
-        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int ResultId { get; set; }
-
-        public string PartyName { get; set; }
-        public string PartyCode { get; set; }
-        public int Votes { get; set; } = -1;
-        public double Percentage { get; set; } = double.NaN;
-    }
-}
