+++ Lavinia-api/Tests/ModelBuilderTests.cs
@@ -24,7 +24,7 @@ namespace LaviniaApi.Tests
             Assert.Equal(556254, dm.Population);
         }
 
-        // Tests a normal input for BuildElectionParameters
+       // Tests a normal input for BuildElectionParameters
         [Fact]
         public static void BuildElectionParametersTest()
         {
@@ -32,11 +32,8 @@ namespace LaviniaApi.Tests
             {
                 new ElectionFormat().Parse("2017;Sainte Laguës (modified);1.4;4.0;1.8;150;19",
                     FieldParser)
-
             };
-            List<CountyDataFormat> countyData = new List<CountyDataFormat> { new CountyDataFormat().Parse("2017;Akershus;4918;556254;16", FieldParser) };
-            IEnumerable<DistrictMetrics> dmList = ModelBuilder.BuildDistrictMetrics(countyData);
-            IEnumerable<ElectionParameters> epList = ModelBuilder.BuildElectionParameters(election, "PE", dmList);
+            IEnumerable<ElectionParameters> epList = ModelBuilder.BuildElectionParameters(election, "PE");
 
             ElectionParameters ep = epList.First();
             Assert.Equal(1.8, ep.AreaFactor);
@@ -50,9 +47,7 @@ namespace LaviniaApi.Tests
             ListElement<double> ap = ep.Algorithm.Parameters.First();
             Assert.Equal("First Divisor", ap.Key);
             Assert.Equal(1.4, ap.Value);
-            ListElement<int>[] seatList = ep.DistrictSeats.ToArray();
-            Assert.Equal("SUM", seatList[0].Key);
-            Assert.Equal(150, seatList[0].Value);
+            Assert.Equal(150, ep.DistrictSeats);
         }
 
         // Tests different DistrictSeat input for BuildElectionParameters
@@ -65,79 +60,9 @@ namespace LaviniaApi.Tests
                 new ElectionFormat().Parse("2017;Sainte Laguës (modified);1.4;4.0;1.8;150;19",
                     FieldParser)
             };
-            List<CountyDataFormat> countyData = new List<CountyDataFormat> { new CountyDataFormat().Parse("2017;Akershus;4918;556254;16", FieldParser) };
-            IEnumerable<DistrictMetrics> dmList = ModelBuilder.BuildDistrictMetrics(countyData);
-            IEnumerable<ElectionParameters> epList = ModelBuilder.BuildElectionParameters(election, "PE", dmList);
+            IEnumerable<ElectionParameters> epList = ModelBuilder.BuildElectionParameters(election, "PE");
 
             ElectionParameters ep = epList.First();
-            ListElement<int>[] seatList = ep.DistrictSeats.ToArray();
-            Assert.Equal("SUM", seatList[0].Key);
-            Assert.Equal(150, seatList[0].Value);
-
-
-            // Tests behaviour for when DistrictSeats are determined before the election
-            election = new List<ElectionFormat>
-            {
-                new ElectionFormat().Parse("1977;Sainte Laguës (modified);1.4;0.0;-1;155;0",
-                    FieldParser)
-            };
-            countyData = new List<CountyDataFormat>
-            {
-                new CountyDataFormat().Parse("1977;Akershus;0;332561;10", FieldParser),
-                new CountyDataFormat().Parse("1977;Aust-Agder;0;81734;4", FieldParser)
-            };
-            dmList = ModelBuilder.BuildDistrictMetrics(countyData);
-            epList = ModelBuilder.BuildElectionParameters(election, "PE", dmList);
-
-            ep = epList.First();
-            seatList = ep.DistrictSeats.ToArray();
-            Assert.Equal("SUM", seatList[0].Key);
-            Assert.Equal(155, seatList[0].Value);
-            Assert.Equal("Akershus", seatList[1].Key);
-            Assert.Equal(10, seatList[1].Value);
-            Assert.Equal("Aust-Agder", seatList[2].Key);
-            Assert.Equal(4, seatList[2].Value);
-        }
-
-        // Tests a normal input for BuildElectionParameters
-        [Fact]
-        public static void BuildElectionParametersTestV3()
-        {
-            List<ElectionFormat> election = new List<ElectionFormat>
-            {
-                new ElectionFormat().Parse("2017;Sainte Laguës (modified);1.4;4.0;1.8;150;19",
-                    FieldParser)
-            };
-            IEnumerable<ElectionParametersV3> epList = ModelBuilder.BuildElectionParametersV3(election, "PE");
-
-            ElectionParametersV3 ep = epList.First();
-            Assert.Equal(1.8, ep.AreaFactor);
-            Assert.Equal(2017, ep.ElectionYear);
-            Assert.Equal(19, ep.LevelingSeats);
-            Assert.Equal(4.0, ep.Threshold);
-            Assert.Equal(0, ep.TotalVotes);
-            Assert.Equal("PE", ep.ElectionType);
-
-            Assert.Equal("Sainte Laguës (modified)", ep.Algorithm.Algorithm);
-            ListElement<double> ap = ep.Algorithm.Parameters.First();
-            Assert.Equal("First Divisor", ap.Key);
-            Assert.Equal(1.4, ap.Value);
-            Assert.Equal(150, ep.DistrictSeats);
-        }
-
-        // Tests different DistrictSeat input for BuildElectionParameters
-        [Fact]
-        public static void BuildElectionParametersDistrictSeatTestV3()
-        {
-            // Tests behaviour for when DistrictSeats are computed
-            List<ElectionFormat> election = new List<ElectionFormat>
-            {
-                new ElectionFormat().Parse("2017;Sainte Laguës (modified);1.4;4.0;1.8;150;19",
-                    FieldParser)
-            };
-            IEnumerable<ElectionParametersV3> epList = ModelBuilder.BuildElectionParametersV3(election, "PE");
-
-            ElectionParametersV3 ep = epList.First();
             Assert.Equal(150, ep.DistrictSeats);
 
 
@@ -147,7 +72,7 @@ namespace LaviniaApi.Tests
                 new ElectionFormat().Parse("1977;Sainte Laguës (modified);1.4;0.0;-1;155;0",
                     FieldParser)
             };
-            epList = ModelBuilder.BuildElectionParametersV3(election, "PE");
+            epList = ModelBuilder.BuildElectionParameters(election, "PE");
 
             ep = epList.First();
             Assert.Equal(155, ep.DistrictSeats);
