+++ Lavinia-api/Utilities/FieldParser.cs
@@ -40,25 +40,7 @@
             return fields;
         }
         
-        /// API V1
-        /// <summary>
-        ///     Parses a string to any of the known algorithms.
-        /// </summary>
-        /// <param name="value">String representation of the algorith</param>
-        /// <param name="field">Name of the field</param>
-        /// <returns>The Algorithm parsed from the string</returns>
-        /// <exception cref="CsvFileFormatException">Is thrown if the string does not match any known algorithm</exception>
-        public Algorithm ParseAlgorithm(string value, string field)
-        {
-            if (!AlgorithmUtilities.TryParse(value, out Algorithm algorithm))
-            {
-                throw new CsvFileFormatException($"The {field} \"{value}\" is not a valid algorithm name.", File, Line);
-            }
 
-            return algorithm;
-        }
-
-        /// API V2
         /// <summary>
         ///     Parses a string to any of the known algorithms.
         /// </summary>
