+++ Lavinia-api/Utilities/ModelBuilder.cs
@@ -7,141 +7,6 @@ namespace LaviniaApi.Utilities
 {
     public static class ModelBuilder
     {
-        // API v1
-        /// <summary>
-        ///     Builds a list of Country object based on a list of CountryFormat objects
-        /// </summary>
-        /// <param name="countries">A list of CountryFormats</param>
-        /// <returns>A list of Countries</returns>
-        public static List<Country> BuildCountries(List<CountryFormat> countries)
-        {
-            List<Country> countryModels = new List<Country>();
-            foreach (CountryFormat country in countries)
-            {
-                Country countryModel = new Country
-                {
-                    CountryCode = country.CountryCode,
-                    InternationalName = country.InternationalName,
-                    ElectionTypes = new List<ElectionType>()
-                };
-                countryModels.Add(countryModel);
-            }
-
-            return countryModels;
-        }
-
-        /// <summary>
-        ///     Takes a list of ElectionFormat and builds a list of Elections
-        /// </summary>
-        /// <param name="elections">A list of ElectionFormat</param>
-        /// <returns>A list of Elections</returns>
-        public static List<Election> BuildElections(List<ElectionFormat> elections)
-        {
-            List<Election> electionModels = new List<Election>();
-            foreach (ElectionFormat election in elections)
-            {
-                Election electionModel = new Election
-                {
-                    Year = election.Year,
-                    Algorithm = election.Algorithm,
-                    FirstDivisor = election.FirstDivisor,
-                    Threshold = election.Threshold,
-                    Seats = election.Seats,
-                    LevelingSeats = election.LevelingSeats,
-                    Counties = new List<County>()
-                };
-                electionModels.Add(electionModel);
-            }
-
-            return electionModels;
-        }
-
-        /// <summary>
-        ///     Takes a list of ElectionTypeFormat and builds a list of ElectionTypes
-        /// </summary>
-        /// <param name="electionTypes">A list of ElectionTypesFormat</param>
-        /// <returns>A list of ElectionTypes</returns>
-        public static List<ElectionType> BuildElectionTypes(List<ElectionTypeFormat> electionTypes)
-        {
-            List<ElectionType> electionTypeModels = new List<ElectionType>();
-            foreach (ElectionTypeFormat electionType in electionTypes)
-            {
-                ElectionType electionTypeModel = new ElectionType
-                {
-                    InternationalName = electionType.InternationalName,
-                    ElectionTypeCode = electionType.ElectionTypeCode,
-                    Elections = new List<Election>()
-                };
-                electionTypeModels.Add(electionTypeModel);
-            }
-
-            return electionTypeModels;
-        }
-
-        /// <summary>
-        ///     Takes a list of ResultFormat and a list of CountyDataFormat, and builds a list of Counties
-        /// </summary>
-        /// <param name="results">A list of ResultFormat</param>
-        /// <param name="countyData">A list of CountyDataFormat</param>
-        /// <returns>A list of Counties</returns>
-        public static List<County> BuildCounties(List<ResultFormat> results, IEnumerable<CountyDataFormat> countyData)
-        {
-            IEnumerable<CountyDataFormat> countyDataFormats = countyData.ToList();
-
-            Dictionary<string, County> countyModels = new Dictionary<string, County>();
-            foreach (ResultFormat resultFormat in results)
-            {
-                if (countyModels.ContainsKey(resultFormat.Fylkenavn))
-                {
-                    continue;
-                }
-
-                try
-                {
-                    CountyDataFormat curCountyData =
-                        countyDataFormats.Single(cD => cD.County.Equals(resultFormat.Fylkenavn));
-
-                    County countyModel = new County
-                    {
-                        Name = resultFormat.Fylkenavn,
-                        Seats = curCountyData.Seats,
-                        Results = new List<Result>()
-                    };
-                    countyModels.Add(resultFormat.Fylkenavn, countyModel);
-                }
-                catch (InvalidOperationException)
-                {
-                    throw new ArgumentException($"Found no region with the name: {resultFormat.Fylkenavn}");
-                }
-            }
-
-            return countyModels.Values.ToList();
-        }
-
-        /// <summary>
-        ///     Takes a list of ResultFormat and builds a list of Results
-        /// </summary>
-        /// <param name="results">A list of ResultFormat</param>
-        /// <returns>A list of Results</returns>
-        public static List<Result> BuildResults(IEnumerable<ResultFormat> results)
-        {
-            List<Result> resultModels = new List<Result>();
-            foreach (ResultFormat result in results)
-            {
-                Result resultModel = new Result
-                {
-                    PartyName = result.Partinavn,
-                    PartyCode = result.Partikode,
-                    Votes = result.AntallStemmerTotalt,
-                    Percentage = result.OppslutningProsentvis
-                };
-                resultModels.Add(resultModel);
-            }
-
-            return resultModels;
-        }
-
-        // API v2
         /// <summary>
         /// Takes a list of CountyDataFormat and builds a list of DistrictMetrics
         /// </summary>
@@ -161,55 +26,9 @@ namespace LaviniaApi.Utilities
         /// <param name="electionType">Election type code</param>
         /// <returns></returns>
         public static IEnumerable<ElectionParameters> BuildElectionParameters(IEnumerable<ElectionFormat> electionData,
-            string electionType, IEnumerable<DistrictMetrics> districtMetrics)
-        {
-            return electionData.Select(data =>
-            {
-                List<ListElement<int>> districtSeats = new List<ListElement<int>> {
-                    new ListElement<int>
-                    {
-                        Key = "SUM",
-                        Value = data.Seats
-                    }
-                };
-
-                if (data.AreaFactor < 0)
-                {
-                    districtSeats.AddRange(
-                        districtMetrics.Where(dM => dM.ElectionYear == data.Year)
-                            .Select(dM => new ListElement<int>
-                            {
-                                Key = dM.District,
-                                Value = dM.Seats
-                            }));
-                }
-
-
-                return new ElectionParameters
-                {
-                    Algorithm = BuildAlgorithmParameters(data),
-                    AreaFactor = data.AreaFactor,
-                    DistrictSeats = districtSeats,
-                    ElectionType = electionType,
-                    ElectionYear = data.Year,
-                    LevelingSeats = data.LevelingSeats,
-                    Threshold = data.Threshold,
-                    TotalVotes = 0
-                };
-            });
-        }
-
-        /// <summary>
-        /// Takes a list of ElectionFormat, an election type and a list of DistrictMetrics, and returns a list of ElectionParameters.
-        /// The returned list contains information about which parameters were used for each election.
-        /// </summary>
-        /// <param name="electionData">List of ElectionFormat</param>
-        /// <param name="electionType">Election type code</param>
-        /// <returns></returns>
-        public static IEnumerable<ElectionParametersV3> BuildElectionParametersV3(IEnumerable<ElectionFormat> electionData,
         string electionType)
         {
-            return electionData.Select(data => new ElectionParametersV3
+            return electionData.Select(data => new ElectionParameters
             {
                 Algorithm = BuildAlgorithmParameters(data),
                 AreaFactor = data.AreaFactor,
